requirements.txt
Twisted==12.2.0
pyOpenSSL==0.13
scrypt==0.5.5
txredisapi==0.9
zope.interface==4.0.1
